Add workspace to safe directory
git-clone task clones our multi-tenancy-gitops-apps repo to a
workspace mapped to a PVC. The owner of the workspace is not set
explicitly by default. This causes issues to when we want to run git
operations on the workspace as a different user. Since, Git v2.35.2,
git operations on directory owned by different use is not allowed. See
https://github.blog/2022-04-12-git-security-vulnerability-announced/,
CVE-2022-24765.

Easiest solution is add the directory as a safe directory. For now,
this will work.

A proper solution would be to use git operation as the same user of
the directory owner, but this might be tricky given the workspace is
mounted from a PVC. Some ideas to try is to set security context
runAsUser and fsGroup, but NSF storage may ignore such settings.